+++ packages/terra-alert/tests/wdio/alert-spec.js
@@ -1,64 +1,61 @@
 Terra.describeViewports('Alert', ['tiny', 'large'], () => {
-  describe('Default', () => {
-    before(() => { browser.url('/#/raw/tests/terra-alert/alert/default-alert'); });
-    Terra.it.validatesElement();
+  it('displays default Alert', () => {
+    browser.url('/#/raw/tests/terra-alert/alert/default-alert');
+    Terra.validates.element('default');
   });
 
-  describe('Types', () => {
-    before(() => { browser.url('/#/raw/tests/terra-alert/alert/alert-type'); });
-    Terra.it.validatesElement();
+  it('displays types of Alert', () => {
+    browser.url('/#/raw/tests/terra-alert/alert/alert-type');
+
+    Terra.validates.element('types');
   });
 
-  describe('Wrapped Content', () => {
-    before(() => { browser.url('/#/raw/tests/terra-alert/alert/alert-with-wrapped-content'); });
-    Terra.it.validatesElement();
+  it('displays Alert with Wrapped Content', () => {
+    browser.url('/#/raw/tests/terra-alert/alert/alert-with-wrapped-content');
+
+    Terra.validates.element('wrapped content');
   });
 
-  describe('Responsive', () => {
-    before(() => { browser.url('/#/raw/tests/terra-alert/alert/responsive-example'); });
-    Terra.it.validatesElement();
+  it('displays Responsive Alert', () => {
+    browser.url('/#/raw/tests/terra-alert/alert/responsive-example');
+
+    Terra.validates.element('responsive');
   });
 
-  describe('Titles', () => {
-    before(() => { browser.url('/#/raw/tests/terra-alert/alert/alert-title'); });
-    Terra.it.validatesElement();
+  it('displays Alert with titles', () => {
+    browser.url('/#/raw/tests/terra-alert/alert/alert-title');
+
+    Terra.validates.element('titles');
   });
 
-  describe('Custom', () => {
-    before(() => { browser.url('/#/raw/tests/terra-alert/alert/custom-alert'); });
-    Terra.it.validatesElement();
+  it('displays custom Alert', () => {
+    browser.url('/#/raw/tests/terra-alert/alert/custom-alert');
+
+    Terra.validates.element('custom');
   });
 
-  describe('Action Button', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-alert/alert/alert-action-button');
-    });
+  it('displays Alert with action button', () => {
+    browser.url('/#/raw/tests/terra-alert/alert/alert-action-button');
 
-    Terra.it.validatesElement();
-
-    it('should be register actions', () => {
-      expect(browser.getText('#actionAlert')).to.have.string('clicked 0 times');
-      browser.click('#actionAlert button');
-      expect(browser.getText('#actionAlert')).to.have.string('clicked 1 times');
-    });
-
-    after(() => browser.moveToObject('#root', 0, 700));
+    Terra.validates.element('action button');
   });
 
-  describe('Dismissible', () => {
-    before(() => {
-      browser.url('/#/raw/tests/terra-alert/alert/alert-dismissible');
-    });
+  it('should register actions', () => {
+    expect(browser.getText('#actionAlert')).to.have.string('clicked 0 times');
+    browser.click('#actionAlert button');
+    expect(browser.getText('#actionAlert')).to.have.string('clicked 1 times');
+  });
 
-    Terra.it.validatesElement();
+  it('displays dismissible Alert', () => {
+    browser.url('/#/raw/tests/terra-alert/alert/alert-dismissible');
 
-    it('should dismiss', () => {
-      browser.click('#dismissibleAlert button');
-      expect(browser.getText('#dismissed')).to.equal('Alert was dismissed');
-    });
+    Terra.validates.element('dismissible');
+  });
 
-    Terra.it.validatesElement('dismissed');
+  it('should dismiss', () => {
+    browser.click('#dismissibleAlert button');
+    expect(browser.getText('#dismissed')).to.equal('Alert was dismissed');
 
-    after(() => browser.moveToObject('#root', 0, 700));
+    Terra.validates.element('dismissed');
   });
 });